data_syncer/pyproject.toml
[tool.poetry]
name = "data-syncer"
version = "0.1.0"
description = ""
authors = ["h3nnn4n <git@h3nnn4n.me>"]
license = "MIT"
readme = "README.md"
packages = [{include = "data_syncer"}]

[tool.poetry.dependencies]
python = "^3.10"
pyserial = "^3.5"
opi-gpio = "^0.5.2"
python-decouple = "^3.7"
influxdb = "^5.3.1"


[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.11"
ipython = "^8.8.0"
mypy = "^1.2.0"
pyright = "^1.1.305"
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py310']

[tool.mypy]
ignore_missing_imports = true

[tool.pyright]
include = ["."]
venv = ".venv"
venvPath = "./"
